Fix UnsupportedOperationException on reuse
Using the ComboAnalyzer through a NamedAnalyzer (which ElasticSearch does),
causes the TokenStreamComponents to be reused without having the
ComboAnalyzer knowing because of the default PerFieldReuseStrategy.

Getting rid of the NeverReuseStrategy for GlobalReuseStrategy.
Making ReusableTokenStreamComponents call the ComboAnalyzer to refresh
itself.
There are no threading issues because the ReuseStrategy use a ThreadLocal,
just like ComboAnalyzer does.

Closes #7 and, #8.